feat: carousel 라이브러리 설치 및 테스트
