+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^4.29.5",
+    "@trendyol-js/react-carousel": "^3.0.2",
     "@types/node": "20.1.0",
     "@types/react": "18.2.6",
     "@types/react-dom": "18.2.4",
